post-checkout: ensure non-interactive merge
Similar to #12, but does not use export and uses "--no-edit" flag to
ensure that editor is not brought up.

Before, failed (sometimes) with the following error:
error: Terminal is dumb, but EDITOR unset
I cannot make out the necessary or sufficient condition to trigger this
error. We had this work a couple times. It broke on a branch which had
merges itself (but would apply cleanly).